requirements.txt
# ProspectIQ Backend Dependencies
# Billing SDK — local path; publish to private PyPI for production
billing-sdk @ file:../digitillis-platform/packages/billing-sdk
stripe>=8.0
anthropic>=0.40.0
httpx>=0.27.0
supabase>=2.10.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
pyyaml>=6.0.2
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
resend>=2.5.0
rich>=13.9.0
typer>=0.15.0
APScheduler>=3.10.0
dnspython>=2.6.0
python-multipart>=0.0.9
scipy>=1.13.0
